Move some orphan instances from Syntax.Concrete.Pretty to Syntax.Common
This fixes a nasty import cycle caused by Agda.Syntax.Internal importing Agda.Syntax.Concrete.Pretty
